Make GroupsScreen functional component, use selectors for searching groups.
